requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
flask-cors==6.0.1
psycopg2-binary==2.9.10
marshmallow==3.20.1
flask-marshmallow==0.15.0
